feat: App.tsx 리팩터링 및 컴포넌트 분리\n\n- CalendarViewPanel, EventFormPanel, EventListPanel 컴포넌트 추가\n- 상태 관리 및 UI 로직을 컴포넌트로 분리하여 가독성 및 유지보수성 향상\n- 기존의 복잡한 UI 구조를 단순화하고 재사용성을 높임
